Revert "Support for Rack middleware"
This reverts commit 0e2fdba5f7cb8cb086be1f63179fbb63ad5feadb.
