+++ lib/hanami/router.rb
@@ -1,37 +1,49 @@
 # frozen_string_literal: true
 
+require "rack/utils"
+
 module Hanami
   # Rack compatible, lightweight and fast HTTP Router.
   #
   # @since 0.1.0
-  #
-  # rubocop:disable Metrics/ClassLength
-  class Router
+  class Router # rubocop:disable Metrics/ClassLength
     require "hanami/router/version"
     require "hanami/router/error"
-    require "hanami/router/inner"
-    require "hanami/router/middleware"
+    require "hanami/router/segment"
+    require "hanami/router/redirect"
+    require "hanami/router/prefix"
+    require "hanami/router/params"
+    require "hanami/router/trie"
 
     def self.define(&blk)
       blk
     end
 
     def initialize(base_url: DEFAULT_BASE_URL, prefix: DEFAULT_PREFIX, resolver: DEFAULT_RESOLVER, &blk)
-      @inner = Inner.new(base_url, prefix, resolver)
-      @stack = Middleware::Stack.new
+      @base_url = base_url
+      # TODO: verify if Prefix can handle both name and path prefix
+      @path_prefix = Prefix.new(prefix)
+      @name_prefix = Prefix.new("")
+      @resolver = resolver
+      @fixed = {}
+      @variable = {}
+      @globbed = {}
+      @mounted = {}
+      @named = {}
       instance_eval(&blk)
-      freeze
-    end
-
-    def freeze
-      @app = @stack.finalize(inner)
-      @app.freeze
-      remove_instance_variable(:@stack)
-      super
     end
 
     def call(env)
-      @app.call(env)
+      endpoint, params = lookup(env)
+
+      unless endpoint
+        return not_allowed(env) ||
+               not_found
+      end
+
+      endpoint.call(
+        _params(env, params)
+      ).to_a
     end
 
     def root(to:)
@@ -76,41 +88,91 @@ module Hanami
     end
 
     def redirect(path, to:, as: nil, code: DEFAULT_REDIRECT_CODE)
-      inner.redirect(path, to: to, as: as, code: code)
+      get(path, to: _redirect(to, code), as: as)
     end
 
     def scope(path, &blk)
-      stack.with(path) do
-        inner.scope(path) do
-          instance_eval(&blk)
-        end
+      path_prefix = @path_prefix
+      name_prefix = @name_prefix
+
+      begin
+        @path_prefix = @path_prefix.join(path.to_s)
+        @name_prefix = @name_prefix.join(path.to_s)
+        instance_eval(&blk)
+      ensure
+        @path_prefix = path_prefix
+        @name_prefix = name_prefix
       end
     end
 
-    def use(middleware, *args, &blk)
-      stack.use(middleware, args, &blk)
-    end
-
     def mount(app, at:, **constraints)
-      inner.mount(app, at: at, **constraints)
+      path = _prefixed_path(at)
+      prefix = Segment.fabricate(path, **constraints)
+      @mounted[prefix] = @resolver.call(path, app)
     end
 
     def path(name, variables = {})
-      inner.path(name, variables)
+      @named.fetch(name.to_sym) do
+        raise InvalidRouteException.new(name)
+      end.expand(:append, variables)
+    rescue Mustermann::ExpandError => exception
+      raise InvalidRouteExpansionException.new(name, exception.message)
     end
 
     def url(name, variables = {})
-      inner.url(name, variables)
+      @base_url + path(name, variables)
     end
 
     def recognize(env, params = {}, options = {})
       require "hanami/router/recognized_route"
       env = env_for(env, params, options)
-      endpoint, params = inner.lookup(env)
+      endpoint, params = lookup(env)
 
       RecognizedRoute.new(
-        endpoint, Params.call(env, params)
+        endpoint, _params(env, params)
       )
+    end
+
+    def fixed(env)
+      @fixed.dig(env["REQUEST_METHOD"], env["PATH_INFO"])
+    end
+
+    def variable(env)
+      @variable[env["REQUEST_METHOD"]]&.find(env["PATH_INFO"])
+    end
+
+    def globbed(env)
+      @globbed[env["REQUEST_METHOD"]]&.each do |path, to|
+        if (match = path.match(env["PATH_INFO"]))
+          return [to, match.named_captures]
+        end
+      end
+
+      nil
+    end
+
+    def mounted(env)
+      @mounted.each do |prefix, app|
+        next unless (match = prefix.peek_match(env["PATH_INFO"]))
+
+        # TODO: ensure compatibility with existing env["SCRIPT_NAME"]
+        # TODO: cleanup this code
+        env["SCRIPT_NAME"] = env["SCRIPT_NAME"].to_s + prefix.to_s
+        env["PATH_INFO"] = env["PATH_INFO"].sub(prefix.to_s, "")
+        env["PATH_INFO"] = "/" if env["PATH_INFO"] == ""
+
+        return [app, match.named_captures]
+      end
+
+      nil
+    end
+
+    def not_allowed(env)
+      (_not_allowed_fixed(env) || _not_allowed_variable(env)) and return NOT_ALLOWED
+    end
+
+    def not_found
+      NOT_FOUND
     end
 
     protected
@@ -153,13 +215,116 @@ module Hanami
     DEFAULT_RESOLVER = ->(_, to) { to }
     DEFAULT_REDIRECT_CODE = 301
 
+    NOT_FOUND = [404, { "Content-Length" => "9" }, ["Not Found"]].freeze
+    NOT_ALLOWED = [405, { "Content-Length" => "11" }, ["Not Allowed"]].freeze
+
+    PARAMS = "router.params"
+    EMPTY_PARAMS = {}.freeze
     EMPTY_RACK_ENV = {}.freeze
 
-    attr_reader :inner, :stack
+    def lookup(env)
+      endpoint = fixed(env)
+      return [endpoint, EMPTY_PARAMS] if endpoint
 
+      variable(env) || globbed(env) || mounted(env)
+    end
+
+    # rubocop:disable Metrics/AbcSize
+    # rubocop:disable Metrics/MethodLength
     def add_route(http_method, path, to, as, constraints)
-      inner.add_route(http_method, path, to, as, constraints)
+      path = _prefixed_path(path)
+      to = @resolver.call(path, to)
+
+      if globbed?(path)
+        @globbed[http_method] ||= []
+        @globbed[http_method] << [Segment.fabricate(path, **constraints), to]
+      elsif variable?(path)
+        @variable[http_method] ||= Trie.new
+        @variable[http_method].add(path, to, constraints)
+      else
+        @fixed[http_method] ||= {}
+        @fixed[http_method][path] = to
+      end
+
+      @named[_prefixed_name(as)] = Segment.fabricate(path, **constraints) if as
+    end
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/AbcSize
+
+    # def add_route(http_method, path, to, as, constraints, &blk)
+    #   (to || blk) or raise "missing endpoint"
+    #   to = Block.new(blk) if to.nil?
+
+    #   path = _prefixed_path(path)
+
+    #   if variable?(path)
+    #     @variable[http_method] ||= Trie.new
+    #     @variable[http_method].add(path, to, constraints)
+    #   else
+    #     @fixed[http_method] ||= {}
+    #     @fixed[http_method][path] = to
+    #   end
+
+    #   # FIXME: pass constraints
+    #   @named[_prefixed_name(as)] = Segment.fabricate(path, {}) if as
+    # end
+
+    def variable?(path)
+      /:/.match?(path)
+    end
+
+    def globbed?(path)
+      /\*/.match?(path)
+    end
+
+    def _prefixed_path(path)
+      @path_prefix.join(path).to_s
+    end
+
+    def _prefixed_name(name)
+      @name_prefix.relative_join(name, "_").to_sym
+    end
+
+    def _redirect(to, code)
+      body = Rack::Utils::HTTP_STATUS_CODES.fetch(code) do
+        raise UnknownHTTPStatusCodeError.new(code)
+      end
+
+      destination = _prefixed_path(to)
+      Redirect.new(destination, ->(*) { [code, { "Location" => destination }, [body]] })
+    end
+
+    def _params(env, params)
+      params ||= {}
+      env[PARAMS] ||= {}
+      env[PARAMS].merge!(Rack::Utils.parse_nested_query(env["QUERY_STRING"]))
+      env[PARAMS].merge!(params)
+      env[PARAMS] = Params.deep_symbolize(env[PARAMS])
+      env
+    end
+
+    def _not_allowed_fixed(env)
+      found = false
+
+      @fixed.each_value do |routes|
+        break if found
+
+        found = routes.key?(env["PATH_INFO"])
+      end
+
+      found
+    end
+
+    def _not_allowed_variable(env)
+      found = false
+
+      @variable.each_value do |routes|
+        break if found
+
+        found = routes.find(env["PATH_INFO"])
+      end
+
+      found
     end
   end
 end
-# rubocop:enable Metrics/ClassLength
